requirements.txt
pytest==7.1.2
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
Appium-Python-Client>=2.3.0
tidevice==0.7.13
selenium~=4.16.0